Bottle Top Race game in your Android!
Available at https://play.google.com/store/apps/details?id=com.lamboratory.android.bottleTop.race

Original control method: Play this racing game by hitting your phone like a bottle top. Click on the screen to point and hit your phone with your finger. The harder you hit, the longer the shot. Watch the video to see it in action.

Multiplayer game in one Android phone. Unlimited number of players.
Try to be the first to get to the finish line to win the race.
Multiple tracks available. Train in the basic circuit or compete with your friends in the complex circuit.
Select the number of laps to complete the race.
Realistic controls, just like playing with a bottle cap.
Apps2SD enabled, you can move it to the SD card. Don't worry about the space it takes.

This is a free game, with no annoying ads. If you like the game rate it in the Android Market and consider making a donation from inside the game with in-app billing ;)

Available in English and Spanish (Carreras de chapas). Contact us if you can translate it to any other language.

Your feedback is welcome. Add your comments in the Android Market or send them to the email address listed there. We want to keep improving our games!

Note: We are not responsible if you hit your phone too hard and break it, or if you hurt your finger :p